tensorflow/requirements.txt
tensorflow==2.3.1 # cuda 10.1
numpy
scikit-image
PyYAML
imgaug